flask_backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
cryptography==43.0.3
Werkzeug==3.0.6
marshmallow==3.22.0
pytest==8.3.3
coverage==7.6.4
gunicorn==23.0.0
PyMySQL==1.1.1
reportlab==4.2.5
